General Fixes (#70)
* remove unavailable function

* update assimp version

* remove unused includes

* initialize variable

* cast variable

* Update DebugMessenger.cpp

* cleanup

* remove unused includes

* cleanup

* add cstring back

+++ source/pipelines/custom/TexturedPlane.cpp
@@ -5,7 +5,6 @@
 #include "../../components/Camera.h"
 #include "../../components/LogicalDevice.h"
 #include "../../components/PhysicalDevice.h"
-#include "../../objects/RenderObject.h"
 #include "../../objects/UniformBuffer.h"
 #include <stdexcept>
 
@@ -86,7 +85,7 @@ void TexturedPlane::createDescriptorSets()
   const VkDescriptorSetAllocateInfo allocateInfo {
     .sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO,
     .descriptorPool = descriptorPool,
-    .descriptorSetCount = static_cast<uint32_t>(logicalDevice->getMaxFramesInFlight()),
+    .descriptorSetCount = logicalDevice->getMaxFramesInFlight(),
     .pSetLayouts = layouts.data()
   };
